🔨[refactor] task response dto 공통 inner class 분리 #109
