+++ src/main/java/com/daon/backend/task/dto/response/FindProjectsResponseDto.java
@@ -1,6 +1,5 @@
 package com.daon.backend.task.dto.response;
 
-import com.daon.backend.task.domain.project.Project;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -21,22 +20,5 @@ public class FindProjectsResponseDto {
         this.workspaceId = workspaceId;
         this.totalCount = projects.size();
         this.projects = projects;
-    }
-
-    @Getter
-    @AllArgsConstructor
-    public static class ProjectSummary {
-
-        private Long projectId;
-
-        private String title;
-
-        private String description;
-
-        public ProjectSummary(Project project) {
-            this.projectId = project.getId();
-            this.title = project.getTitle();
-            this.description = project.getDescription();
-        }
     }
 }
