+++ src/main/java/com/daon/backend/task/dto/response/WorkspaceListResponseDto.java
@@ -1,7 +1,5 @@
 package com.daon.backend.task.dto.response;
 
-import com.daon.backend.task.domain.workspace.Workspace;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 import java.util.List;
@@ -16,28 +14,5 @@ public class WorkspaceListResponseDto {
     public WorkspaceListResponseDto(List<WorkspaceSummary> workspaces) {
         this.totalCount = workspaces.size();
         this.workspaces = workspaces;
-    }
-
-    @Getter
-    @AllArgsConstructor
-    public static class WorkspaceSummary {
-
-        private Long workspaceId;
-
-        private String title;
-
-        private String imageUrl;
-
-        private String division;
-
-        private String description;
-
-        public WorkspaceSummary(Workspace workspace) {
-            this.workspaceId = workspace.getId();
-            this.title = workspace.getTitle();
-            this.imageUrl = workspace.getImageUrl();
-            this.division = workspace.getDivision().name();
-            this.description = workspace.getDescription();
-        }
     }
 }
